requirements-2.txt
Flask
gunicorn
requests
requests-oauthlib
python-dotenv
tweepy
